requirements.txt
feedparser==6.0.11
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
schedule==1.2.1
instagrapi==2.1.0
deep-translator==1.11.4
langdetect==1.0.9
